+++ test/components/Output/Output.spec.js
@@ -9,6 +9,8 @@ import React from 'react';
 
 import { fireEvent, render, waitFor } from '@testing-library/react';
 
+import { TooltipProvider } from '@camunda/design-system';
+
 import Output, { getWaitingContext } from '../../../lib/components/Output/Output';
 
 import {
@@ -935,9 +937,11 @@ const Wrapper = (props) => {
   const pluginsProviderValue = usePluginsProviderValue();
 
   return (
-    <PluginContext.Provider value={ pluginsProviderValue }>
-      { props.children }
-    </PluginContext.Provider>
+    <TooltipProvider>
+      <PluginContext.Provider value={ pluginsProviderValue }>
+        { props.children }
+      </PluginContext.Provider>
+    </TooltipProvider>
   );
 };