+++ webpack.config.js
@@ -19,10 +19,8 @@ module.exports = {
     'react/jsx-dev-runtime': 'react/jsx-dev-runtime',
     'react-dom': 'react-dom',
     'react-dom/server': 'react-dom/server',
-    '@carbon/react': '@carbon/react',
-    '@carbon/icons-react': '@carbon/icons-react',
     '@camunda/design-system': '@camunda/design-system',
-    'lucide-react': 'lucide-react'
+    '@camunda/design-system/icons': '@camunda/design-system/icons'
   },
   resolve: {
     extensions: [ '.js', '.jsx' ]